Fixing outdated/deprecated code that prevented compilation
The distinction between Strings and Bytes was made in OCaml 4.02, which
means that this code is only compatible with OCaml 4.02 or higher.

It doesn't matter nowadays because the code runs.

The only thing that doesn't work are the Xorg commands. Sadly, the
demos don't work anymore because all of required Xorg.

What is impressive is that this 20+ year old code required just minor
adjustments to be compilable, and potentially, still work in today's
world. This is a testament to write code using the official stable
libraries of any language. The same, I've heard, happens with Haskell
old code written using only the Prelude.